requirements.txt
langchain==0.3.7
langchain-openai==0.2.9
langgraph==0.2.45
python-dotenv==1.0.0